Make it possible to not add requests to 'waiting' queue in poolboy
When `:poolboy.transaction/3` is used, poolboy puts the request into a
'waiting' queue if all workers are busy. The caller is then blocked until a
worker becomes available and the request is processed (or passed timeout value
is reached). This is helpful during short load spikes, however if the overload
is sustained, the queue can become too expensive to manage: poolboy uses
`:queue.filter/2` to remove items from the queue, which is an O(n) operation.

It may be preferable to not add a request to the queue and abandon it
immediately if all workers are busy to avoid overloading the pool and ensuring
that new requests that come in don't have to wait for the ever-growing queue of
previous requests is processed.

This commit adds a configuration option for disabling adding a request to the
'waiting' queue: poolboy allows to do that when `:poolboy.checkout/3` is used
instead of `:poolboy.transaction/3`. The default value for the option is
`true`, which is the current behaviour.