sanitize transport options
*Problem*

Some transport options are invalid when passed to the tcp transport,
but machine_gun pools are specified not for a specific host-port pair
but for a so called "pool group", this means that options set for the
ssl transport are being passed to the tcp transport and causing
failure.

*Solution*

Drop invalid options depending on the given transport.
Currently only `verify`, as that is the option that is being specified.